Revised TrackerActivity and BackgroundTrackerService  to support tracking when network service is unavailable or intermittent.
The model used is simple.  Each point received from the location service is saved to the phone's database.  It is then passed to an Async thread which handles communication with the server.  The thread sleeps -- timed wait -- while waiting for network connectivity to return.

Known issues:  The UI gets a little sluggish when too many threads are waiting because of the lack of connectivity. (Too many is 150 or so points.)   Some few points are lost if network connectivity is lost between the time a thread checks for (and finds) connectivity and the successful completion of the Http Post.

Also made changes to the Tracker interface, using buttons instead of menus.  Changes to other sources are minimal and cosmetic.