Refactored and modified Tracker code to improve its over functionality and performance.  In this version points gather from the Location Listener are first saved to the Db and then sent to the server only if there is service (MOBILE or WIFI) at the time when the point is gathered.  This fixes the issue of missing points.  When the expedition is finished, any points that were not sent to the server can be synced by the user -- a button for that purpose is displayed for expeditions whose total points exceed their syced points.
Also, improved the responsiveness of the user interface (UI) by moving the Location Listener to an Async thread.  There is still room for improvement here -- possibly by a better algorithm -- but things are much improved.

This version of Tracker works even if there is no service when Tracker is started. In this case, it creates a random Expedition ID that is used during tracking.  When the completed track is later displayed, POSIT detects that it has used a random ID and syncs with the server (provided there is WIFI or MOBILE service).

This changes needs additional testing.

Known issue:  I disabled the "Notification" function in Tracker.  The problem is that when Tracker is started from the Notification status bar (after POSIT has been stopped) it makes it impossible to start PositMain, either from the Application menu or by hitting the back key when in TrackerActivity.  It seems that  when started by the Notification as  PendingIntent, TrackerActivity becomes the Main activity and thereafter it is started whenever POSIT is started -- I know that doesn't make sense but that how it seems.