Rewrote parts of TrackerBackgroundService to improve its design and functionality.  It now uses only a single Async Task for sending points to the Posit Server.  The design is based on this nice little tutorial:  http://github.com/commonsguy/cw-android/tree/master/Service/WeatherPlus/.
Also made some minor improvements to the interface, including updating the Expedition Id immediately upon receiving it from the Server.

Known issues: there are still issues with responsiveness and with excessive garbage collection in TrackerActivity after the Trackers goes from running to idle.  These seem to be associated with the Google map.