Integrates into POSIT Prasanna's Adhoc package (org.hfoss.adhoc), which relies entirely on UDP Sockets (rather than raw sockets)  for peer-to-peer communication.
The current version removes (i.e.temporarily comments out) the RWG Service and incorporates a simple adhoc broadcast service.  The broadcast service can be turned on/off in PositMain (as was the case with RWG).  When adhoc is on, Finds are broadcast through UDP sockets to other phones that are in adhoc mode and listening for packets marked RWG.   This means that a new Find will reach all reachable phones at the time when it is saved on the phone that found it.  It will not later be transmitted to phones that were not reachable when it was found--that is the additional  functionality that RWG provided.

This is the first step towards a re-implementation of RWG within the application (Java) layer.  It removes the dependency on raw sockets and also removes all reliance on C binaries. So this version should run (although it has not yet been tested on) off-the-shelf (unrooted) phones.

The next step is to implement the RWG protocol on top of this adhoc package.

Perhaps, for backward compatibility purposes, we should create a separate branch for Chris Fei's implementation of RWG??