Fixes issue #82 -- Tracker gets stuck in SYNCING_POINTS mode.    Following completion of the Async loop in SendExpeditionPointTask-- the loop that processes each of the unsynced points -- the Tracker state is now set to IDLE.   This means that the Tracker will be in the iDLE state even if some of the processed points were not successfully synced. If the network connection is lost during syncing, it may be that 1 (or more??) points will be left unsynced.  This solution leaves those points to be synced (by user selection) at some later time.
One side effect of this solution:  If the network is lost while syncing unsynced points, the Tracker will remain in the SYNCING_POINTS state until the network becomes available and it can resume processing its points.   It's not clear whether this should be changed.  Perhaps a better solution might be if the network is not back up within, say, 5 minutes, the Async thread is stopped and the Tracker is put back in the IDLE state??