Incremental changes to Tracker activity.  Fixes  a bug in TrackerBackgroundService that was causing changes to the MINIMUM_DISTANCE setting to be ignored by the Location Listener.   Also, changed the MINIMUM_RECORDING_DISTANCE from 0 to 2000 (millisecs). This doesn's seem to have an effect on the  number of points gathered, which seems to be controlled by the MINIMUM_DISTANCE setting.
Also modifed the TrackerActivity to remove TrackerOverlays from mapView before updating with a new track.  These updates happen very frequently and this bug was probably causes a degradation in responsiveness to the MapView as Overlays were being piled up on the View.

There are still some flakiness issues with the Tracker's MapView.