Implements Ryan's patch for SMS messaging plugin.  The plugin currently attaches to FindActivity menu.  It starts an activity that lets the user input a message and phone number and it transmits that message.   It seems to be independent of other plugins, although I only tested it with the derived Clp plugin.
TODO:  This needs to be integrated into POSIT.  A next step would be to allow the phone number to be set in an SMS Settings preference and to allow the currently selected Find to be sent.  Perhaps the menu option should only be enabled for an already-saved Find -- i.e., user can't SMS a new find until it has been saved.