Implements a CsvFind plugin.  A CsvFind is a Find that is constructed from a CSV row -- i.e., a comma-separated string of fields.  The CSV file must begin with a header row that begins with #GUID, COL1, COL2, ....  The column names (beside Guid) are used as the Find's field names.  The columns in the CSV must not contain embedded commas.
There are two ways to use this plugin.  By default, the CsvFind plugin comes with NYC Museums data downloaded from http://nycopendata.socrata.com/.  If this plugin is true, then Posit becomes an guide to NYC museums, linking to phone numbers, maps and directions, web sites.   If CsvFind is false and BasicFind is true, then the Csv files can be read using the FileViewActivity under the ViewFinds menu.   In this version it might only work with the Museum data, but in future versions it should, in principle, be able to open any Csv file.

TODO:  Additional work needs to be done to generalize this so it will work with different Csv files.