Partially addresses issue #330 -- sets the summary of the Locale preference.
This patch successfully adds a summary to the Locale preference in the Settings activity.  However, the value of the summary appears to be set from the string resource, rather than from the current value of the preference.  So it always shows "English" as the summary when you open Settings, even if the current locale is France.  This should be fixed in a second patch.